Measure is implemented but not working yet.
Need to debug. Right now we need to figure out the mutability issue
with the lattice in the winding number measure. The trait measure
does not require a mutable lattice (makes sense) but the function
get_link_from_point returns a mutable link and requires a mutable
lattice. Should I make a different version of the get link function.