Fixing and adding stuff for getting vertex from lattice.
Im going to need to be able to get Vertex objects for vertices off of the primary sublattice.
This functionality did not exist in the Lattice structure so Im building it now becuase
the clustering needs to be able to hop onto arbritrary points and get the directions at
those verticies regardless of sublattice.